feat: :art: structure the application tables to be a superclass with two subclasses and update imports form this file
a super class for applications with subclasses assistant application and team application

vf-247
