+++ db/tables/team.ts
@@ -1,6 +1,6 @@
 import { departmentsTable } from "@db/tables/departments";
 import mainSchema from "@db/tables/schema";
-import { teamApplicationsTable } from "@db/tables/teamApplication";
+import { teamApplicationsTable } from "@db/tables/applications";
 import { relations } from "drizzle-orm";
 import { boolean, date, serial, text } from "drizzle-orm/pg-core";
 import { integer } from "drizzle-orm/pg-core";
